Move UserConfig to CLI.Config
